+++ tools/CreateMappings.ps1
@@ -16,7 +16,7 @@ $results = @{};
 $warnings = @();
 
 # Find all cmdlet names by help file names in the repository.
-$cmdlets = Get-ChildItem $RootPath -Recurse | Where-Object { $_.FullName -cmatch ".*\\help\\.*-.*.md" -and ($_.FullName -notlike "*Stack*" -or $_.FullName -like "*StackEdge*") };
+$cmdlets = Get-ChildItem $RootPath -Recurse | Where-Object { $_.FullName -cmatch ".*\\help\\.*-.*.md" -and ($_.FullName -notlike "*Stack*" -or $_.FullName -like "*StackEdge*" -or $_.FullName -like "*StackHCI*") };
 
 $k = 0;
 $cmdlets | ForEach-Object {