+++ src/Resources/Resources/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* added properties "Condition", "ConditionVersion" and "Description" to New-AzRoleAssignment
+    - this includes all the relevant changes to the data models
 
 ## Version 2.3.0
 * Updated `Save-AzResourceGroupDeploymentTemplate` to use the SDK.